requirements.txt
streamlit
plotly
pandas
openpyxl
gspread
google-auth
google-auth-oauthlib
google-api-python-client
